pull messagepool into separate package
