+++ chain/messagepool/messagepool_test.go
@@ -1,4 +1,4 @@
-package chain
+package messagepool
 
 import (
 	"testing"
